We need to use our custom function here to get the right return value type
